Extract dictation socket protocol helpers into their own module
Moves out of useAudioDictationRecorder:

- AUDIO_DICTATION_SOCKET_OPEN_TIMEOUT_MS / FRAME_TIMEOUT_MS
- AudioDictationSocketFrame discriminated-union type
- createAudioDictationWebSocketUrl
- sendAudioDictationControlFrame
- waitForSocketOpen
- waitForAudioDictationFrame

These are protocol-only concerns (URL construction, JSON-frame
encoding, listener-based promise helpers) with no React, no audio
pipeline, and no hook state. Lives in a sibling
audio-dictation-protocol module. Another ~110 lines out of the
recorder. Zero behavior change.
